+++ cf/auxiliarycoordinate.py
@@ -1,5 +1,7 @@
 import cfdm
 
+from . import Bounds
+
 from . import mixin
 
 
@@ -41,6 +43,13 @@ class AuxiliaryCoordinate(mixin.Coordinate,
     `nc_clear_variable_groups` and `nc_set_variable_groups` methods.
 
     '''
+    def __new__(cls, *args, **kwargs):
+        '''
+        '''
+        instance = super().__new__(cls)
+        instance._Bounds = Bounds
+        return instance
+
     def __repr__(self):
         '''Called by the `repr` built-in function.